Augmentation de la taille de nb_keys (uint16 -> uint32)

+++ inc/c1.h
@@ -11,7 +11,7 @@ bool test_char_on_str(byte c, uint8 pos, uint8 key_length, byte str[],
 void print_chars(byte chars[], uint8 nb_chars);
 void extract_keys_loop(uint8 i, uint16 *i_key, byte **keys, byte **key_chars,
                        byte key_chars_length[], byte temp_key[], uint8 key_length);
-byte **extract_keys(byte **key_chars, uint16 *nb_keys, uint8 key_chars_length[],
+byte **extract_keys(byte **key_chars, uint32 *nb_keys, uint8 key_chars_length[],
                     uint8 key_length);
 byte **c1(byte str[], uint32 str_length, uint8 key_length,
           uint8 key_chars_length[]);
